requirements.txt
fastapi
uvicorn
python-dotenv
creditagricole_particuliers
pandas
openpyxl
python-multipart
pydantic
lxml
requests